django_react/requirements.txt
# terminal to see what is need => "pip freeze"
Django==3.0.7
#psycopg2==2.2.2
asgiref==3.2
backcall==0.2.0
decorator==4.4.2
django-cors-headers==3.4.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.1
mysqlclient==1.4.6
numpy==1.18.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
# pkg-resources==0.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.4
asgi-redis==1.4.3
channels==2.4.0
channels-redis==2.4.2
django_redis==4.12.1
python-memcached==1.59